Refactor TimeEntryBlock to extract different class for gap time entries (win)
