+++ src/ui/windows/TogglDesktop/TogglDesktop/TogglDesktop.csproj
@@ -331,6 +331,7 @@
     <Compile Include="ui\ViewModels\TimeEntryCellViewModel.cs" />
     <Compile Include="ui\ViewModels\TimeEntryLabelViewModel.cs" />
     <Compile Include="ui\ViewModels\TimeEntryListViewModel.cs" />
+    <Compile Include="ui\ViewModels\TimelineBlockViewModel.cs" />
     <Compile Include="ui\ViewModels\TimelineViewModel.cs" />
     <Compile Include="ui\ViewModels\TimerEntryListViewViewModel.cs" />
     <Compile Include="ui\ViewModels\TimerViewModel.cs" />
